Add an initial hpack scaffold
This patch sets up a [Stack][stack] build-system using [hpack][hpack],
including a README and a copy of the 3-clause BSD license, and builds
a simple executable with the `UnicodeSyntax` language extension.  This
serves as a minimal-viable project template that can be built and
verified in a continuous integration system.  At this point, the
repository contains no code that is particular to Minimalist Grammars,
but rather only the plumbing that we need in order to add MG-specific
code.

A Stack build system consists of a handful of files that describe the
package structure as well as its build dependencies.  First is the
`stack.yaml` file that specifies what packages to build, what
dependency packages are needed, and where to download dependency
packages from.  We use the latest LTS from Stackage, 19.5 with GHC
9.0.2, as our resolver for packages.  If we need additional packages
that are not listed in Stackage (such as repositories on GitHub), we
will need to add a description of how to download and build these
packages.

Then, we need to teach Stack how to build our current package, once it
has sorted out all the dependencies.  This is where [Stack’s hpack
support][stack-hpack] comes in.  Hpack uses a YAML file that describes
project metadata and the build targets in the project—in our case,
just a single executable.  Hpack then generates a Cabal file that
Stack uses to build our executable.  _N.b._ because the Cabal file is
generated, it goes in our `.gitignore`.  Our hpack configuration is
mostly simple, except for the following: for me, it is crucial to
enable the language extension `UnicodeSyntax`, which will allow us to
use unicode characters in our code, such as `∷` in place of `::` and
`→` in place of `->`.

In addition, Hpack lets us point to documentation files, like the
README, as well as the license text, for which we use a 3-clause BSD
license, which is common in the Haskell community.  Although in
general I prefer to license code under the AGPL, the LGPL, or the GPL,
in this case making the license match that of the better part of the
Haskell ecosystem is more important.

Finally, to upload our package to Hackage, we need a `Setup.hs` file
that can contain build hooks, but whose content in our project is
standard boilerplate.

In addition, this patch adds a pretty standard `.gitignore` file for
Haskell, to prevent us from accidentally staging a generated file in
our repository.

[stack]:       https://haskellstack.org/
[hpack]:       https://github.com/sol/hpack
[stack-hpack]: https://docs.haskellstack.org/en/stable/faq/#does-stack-support-the-hpack-specification

Testing
-------

The changes in this patch can be verified by running `stack build &&
stack run`, and verifying that the Old English rune _Ȳr_ “ᚣ” is
printed to the screen.  This verifies that the executable is properly
built by the build script.

License: BSD-3-Clause
Tested-by: Patrick M. Niedzielski <patrick@pniedzielski.net>
Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>