docs: Add a README and license text
This initial patch adds a README and copies of the two licenses that
will be used in this project: GPLv3 (for application code, tests, and
benchmarking), and LGPLv3 (for library code that we expect others to
link to.

Testing
-------------------------------------------------------------------------------

No validation needed to be performed on this patch, as it contains
only Markdown files for documentation.

License: GPLv3+
Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>