plumbing: Add an initial hpack scaffold
This patch sets up a [Stack][stack] build-system using [hpack][hpack],
and builds a simple executable with the `UnicodeSyntax` language
extension.

A Stack build system consists first of a `stack.yaml` file that
specifies what packages to build, what dependency packages are needed,
and where to download dependency packages from.  We use the latest LTS
snapshot from Stackage as our resolver for packages.

Then, we need to teach Stack how to build our current package, once it
has sorted out all the dependencies.  This is where [Stack’s hpack
support][stack-hpack] comes in.  Hpack uses a YAML file that describes
project metadata and the build targets in the project—in our case,
just a single executable.  It is crucial for us to enable the language
extension `UnicodeSyntax`, which will allow us to use unicode
characters in our code, such as `∷` in place of `::` and `→` in place
of `->`.  Hpack then generates a Cabal file that Stack uses to build
our executable.  _N.b._ because the Cabal file is generated, it goes
in our `.gitignore`.

Finally, to upload our package to Hackage, we need a `Setup.hs` file
that can contain build hooks, but whose content in our project is
standard boilerplate.

This patch also adds a pretty standard `.gitignore` file for Haskell,
to prevent us from accidentally staging a generated file in our
repository.

[stack]:       https://haskellstack.org/
[hpack]:       https://github.com/sol/hpack
[stack-hpack]: https://docs.haskellstack.org/en/stable/faq/#does-stack-support-the-hpack-specification

Testing
-------------------------------------------------------------------------------

The changes in this patch can be verified by running `stack build &&
stack run`, and verifying that the Old English rune Ȳr “ᚣ” is printed
to the screen.  This verifies that the executable is properly built by
the build script.

License: GPLv3+
Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>