Add test harness
This patch adds an empty harness for running tests that verify the
correctness of the `mgs` package.  This harness is built using
[tasty][tasty], a framework that integrates [HUnit][hunit],
[QuickCheck][quickcheck], [SmallCheck][smallcheck], and others.

To run tests, execute the command

    stack build --haddock --test

The harness added by this patch currently does nothing except define
an empty test group called “Tests”.  To add a new test for a code
module, we will create a submodule named `Tests`; this module should
expose a `TestTree` that verifies the associated code module.

[tasty]: https://github.com/feuerbach/tasty
[hunit]: https://github.com/hspec/HUnit
[quickcheck]: https://github.com/nick8325/quickcheck
[smallcheck]: https://github.com/strake/smallcheck.hs

Testing
-------

This patch does not change the behavior of the package, except to add
an empty test harness.  The test harness was manually verified by
running the command `stack build --test` and ensuring that it
completes successfully, after running no tests.

License: BSD-3-Clause
Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>
Tested-by: Patrick M. Niedzielski <patrick@pniedzielski.net>