Refactor formatting functions for less duplication
The previous patch implemented the `formatBlahWithFmt` functions in
terms of the corresponding `formatBlah` functions, utilizing the
objects’ `Functor` properties.  This resulted in several nested
`fmap`s which make the code harder to understand at a glance.  This
patch replaces the pairs of functions implementations so that the
`formatBlahWithFmt` functions have the real implementations, and the
`formatBlah` forward directly to the ‘formatBlahWithFmt‘ functions.