Teach Travis CI to build our package
This patch adds a simple configuration from Travis CI, taken directly
from the Stack documentation, which builds and tests our package when
it is pushed to GitHub.  [Travis CI][travis-ci] is a continuous
integration system that can build repositories on GitHub every time
new commits are pushed.  It can be configured to merely build and test
software in the repository in a virtual environment, and to
additionally upload build artifacts to an external service.

Our configuration [is taken directly from][stack] the Stack GitHub
repository, which builds our package and its API documentation, and
runs any tests we have associated with the package.  Because we use
Stack as our build system, the configuration downloads the latest
version of Stack, builds our packages’ dependencies, and builds our
package, its API documentation, and runs its tests.

One of the features of Travis CI is that it provides SVG badges that
track whether particular branches successfully build or not.  As part
of our configuration of Travis CI, we include one such badge in our
README, tracking the master branch of our repository.

[travis-ci]: https://www.travis-ci.com/
[stack]:     https://github.com/commercialhaskell/stack/blob/master/doc/travis-simple.yml
[license]:   https://github.com/commercialhaskell/stack/blob/master/LICENSE

Testing
-------

This patch configures an external build system that only runs when a
commit is pushed, so this patch can only be tested after it is
committed and pushed.

License
-------

Because this configuration is taken from another source, it is
important that we have a license for it.  According to the [repository
this documentation lives][license], it is licensed under the same
3-Clause BSD license we use in this repository, with copyright held by
contributors to Stack.

License: BSD-3-Clause
Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>