Add initial MG module
At the moment, we have no MG specific code in our package.  While we
will eventually add modules for specific MG-related functionality, it
is useful to have an overarching module that re-exports the most
common MG functionality, to avoid users needing to import each small
module they need to use.  This patch adds such a module, which is at
the moment devoid of exports.

The module this patch adds is named `FormalLanguage.MG`.  This name
mirrors that of another package that already exists, for working with
CFGs, which lives under the `FormalLanguage` namespace.  At the
moment, the module consists of documentation describing what a
Minimalist Grammar is, and what the purpose of this module is, along
with a short bibliography of references.

In order for the module to be compiled, we need to add a library to
our Hpack configuration, pointing to subdirectory where the library is
contained.  Because the module doesn’t export anything at the moment,
this patch does not yet add imports to the executable or tests for the
module, which would not have any use.

Testing
-------

This patch does not introduce any functionality, but it does change
the build system.  After applying the patch, it was verified that the
package still builds without warnings by running:

    stack build --test --haddock

License: BSD-3-Clause
Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>
Tested-by: Patrick M. Niedzielski <patrick@pniedzielski.net>