Add `Feature f` type
The feature calculus of Minimalist Grammars is what drives the
structure of derivations according to the grammar.  As such, having a
type that represents the types of role-annotated features, as well
which features match to drive structure-building operations.

This patch adds the `FormalLanguage.MG.Feature` module, which contains
a `Feature f` type constructor for building the role-annotated
features of Simplified Minimalist Grammars from a set of basic
features `f`, as well as basic functions for working with
role-annotated features.  In addition, because role-annotated features
are so fundamental to the construction of Minimalist Grammars, we
export this new module from the main `FormalLanguage.MG` module.

Testing
-------

This patch adds a module test to our test harness, which tests each of
the exported functions and the handwritten typeclass instances.  The
correctness of this patch can be tested by running

    stack build --test --haddock

and then ensuring that all tests pass, and that the resulting Haddock
documentation is complete.

License: BSD-3-Clause
Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>