Refactor `FormalLanguage.MG.Feature.Tests`
This patch refactors the structure of tests for the `Feature` type to
separate out those tests that are particular to the laws of
typeclasses it implements.  This will allow us to write the tests for
typeclass laws more generally (and use the package
[tasty-quickcheck-laws][tql], which implements many such tests for
us).

[tql]: https://hackage.haskell.org/package/tasty-quickcheck-0.10.2/

Testing
-------

Because this patch only changes the structure of existing tests, it
should not affect the correctness of the actual library.  As such,
this patch can be verified by running

    stack build --test

ensuring that all the tests pass, and that all tests from the previous
patch still run.

License: BSD-3-Clause
Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>