Use `tasty-quickcheck-laws` to test `Eq` laws
Code coverage reports generated by `stack test --coverage` showed that
our tests weren’t exercising all the instances of our types.  Although
the particular instances we weren’t testing are the ones that are
automatically derived by the compiler, it is not guaranteed that we
won’t ever rewrite these ourselves.  These tests will look
substantially the same for each type that implements these
typeclasses, though, so it makes sense to write these tests
generically and stamp out new tests for each type that implements a
particular typeclass.

This patch accomplishes this by adding a dependency on the
`tasty-quickcheck-laws` package, which includes generic
QuickCheck-style tests for several common typeclasses, and adding `Eq`
tests for each of the three types in the `FormalLanguage.MG.Feature`
module.

Testing
-------

This patch adds more tests to the `FormalLanguage.MG.Feature.Tests`
module, which can be verified by running

    stack build --test --haddock

and checking that the `Eq Laws` for each of the types pass.

License: BSD-3-Clause
Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>