Make tests for `Ord` laws generic
This patch continues the work of the previous patches in reducing
hard-coded tests for checking typeclass laws.  While the previous
patch did this by including the `tasty-quickcheck-laws` library so we
could use pre-existing tests for `Eq` instances, this patch follows
the same structure in implementing a `Test.Tasty.QuickCheck.Laws.Ord`
module internally for testing `Ord` laws.  This module could hopefully
be cleaned up eventually for inclusion in the `tasty-quickcheck-laws`
library, but for the moment we use it to test each of the `Ord`
instances in the `FormalLanguage.MG.Feature` module.

In the case of the functor type `Feature`, we need to be
careful to only check instances where the internal type has a small
cardinality, so we can effectively test the Antisymmetry Law:
otherwise, we are unlikely to generate enough test cases for `a ≤ b`
and `b ≤ a`.  We thus test only `Feature ()` and `Feature Bool`, not
larger types like `Feature String`.  This should be sufficient to
check the laws.

Testing
-------

This patch adds more tests to the `FormalLanguage.MG.Feature.Tests`
module, which can be verified by running

    stack build --test --haddock

and checking that the `Ord Laws` for each of the types pass.

License: BSD-3-Clause
Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>