Make tests that check `Ord` and `Bounded` interaction generic
This is the third in a series of patches that test laws for typeclass
instances.  However, unlike the previous two, the tests that this
patch implements are not written as necessary laws for an instance to
implement, but rather as laws that any type that is an instance both
of `Ord` and `Bounded` *ought* to implement.  Specifically, this patch
implements a `Test.Tasty.QuickCheck.Laws.BoundedOrd` module, which
checks that `minBound ≤ maxBound`; that for all `a`, `a ≤ maxBound`,
and that for all `a`, `minBound ≤ a`.  None of these laws are
required, but they match the intention of `Bounded` types that are
totally ordered.

Testing
-------

This patch adds more tests to the `FormalLanguage.MG.Feature.Tests`
module, which can be verified by running

    stack build --test --haddock

and checking that the `Bounded and Ord Laws` for each of the types
pass.

License: BSD-3-Clause
Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>