Add dependency on `tasty-travis`
The `tasty-travis` package formats test output in a way that allows
Travis CI to group, fold, and colorize it, while falling back to the
previous formatting for tests run anywhere outside Travis CI.  This
patch adds a dependency on this package and switches the default Tasty
test reporter to this package‘s `travisTestReporter`.

Testing
-------

This patch should not change the tests that are executed at all, and
shouldn’t change the format the test harness’s output either in the
default case.  To check this, run

    stack build --test

and ensure the output is exactly the same as it is in the previous
patch.  To amke sure that Travis CI will use the new output format, we
run the same command, with the `TRAVIS` environment variable set to
true, and make sure that the output uses the Travis CI formatting.

    TRAVIS=true stack build --test

License: BSD-3-Clause
Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>